refactor: restaurant에 종속된 컴포넌트 파일 이름 변경
